Upgrade to 2.0.0 with Resampling Changes
Revised resampling to have behavior in accordance with the teaching style of the course. Fixed several looming bugs relevant to the funciton, namely issues with integer ndarrays and interpolation, and the inability to use ndarrays with ndim=3. Now import all of pyplot, per request of Motaz and Jinsol. Modified setup.py to indicate the precise dependency of Python 3.5.2+.